Add files via upload
This program is a simple voting eligibility system. It asks the user to enter their age. If the age is 18 or above, the program shows that the person is eligible to vote. If the age is less than 18, it shows that the person is not eligible to vote. This program helps beginners understand if–else conditions and user input in C++.